+++ package/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.1-preview] - 2018-10-20
+- Reduced object duplication for scene asset bundles
+- Fixed object order for scene asset bundles not being deterministic
+- Fixed Shader Stripping values from Graphics Settings not applying to built data
+- Fixed various code warnings for Unity 2018.3 and newer version
+- Added forcing asset save on build
+
 ## [1.1.0-preview] - 2018-10-01
 - Fixed an issue where a string hash was being used instead of a file hash causing data to not rebuild